+++ src/main/java/net/lsmith946/adventofcode/problems2020/days/Day3.java
@@ -41,7 +41,7 @@ public class Day3 implements Puzzle {
     }
 
     @Override
-    public int solvePartOne() {
+    public long solvePartOne() {
         int treesHit = countTreesOnRoute(3, 1);
         System.out.println("Number of trees hit: " + treesHit);
         return treesHit;
